+++ docker/freebsd.sh
@@ -107,7 +107,9 @@ main() {
     # clean up
     popd
 
-    apt-get purge --auto-remove -y ${purge_list[@]}
+    if (( ${#purge_list[@]} )); then
+      apt-get purge --auto-remove -y ${purge_list[@]}
+    fi
 
     rm -rf $td
     rm $0
