+++ docker/musl.sh
@@ -48,8 +48,9 @@ main() {
         OUTPUT=/usr/local/ \
         "$@"
 
-    # clean up
-    apt-get purge --auto-remove -y ${purge_list[@]}
+    if (( ${#purge_list[@]} )); then
+      apt-get purge --auto-remove -y ${purge_list[@]}
+    fi
 
     popd
